fix: [api] usage of ProjectManagerListener, which is a deprecated api scheduled for removal
